DB_ESP/event_management/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
Django==5.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Flask==3.1.0
google==3.0.0
googlesearch-python==1.2.4
h11==0.14.0
httpie==3.2.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
lib==4.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
selenium==4.21.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
Werkzeug==3.1.3
wsproto==1.2.0
